chore: refactoring case file reducer
cleaned up the case file reducer and moved selectors and thunks to separate files for better organization
