+++ frontend/src/app/components/containers/complaints/outcomes/supplemental-notes/supplemental-notes-input.tsx
@@ -9,7 +9,7 @@ import { useAppDispatch } from "../../../../../hooks/hooks";
 import { openModal } from "../../../../../store/reducers/app";
 import { CANCEL_CONFIRM } from "../../../../../types/modal/modal-types";
 import { MAX_CHARACTERS } from "../../../../../constants/general";
-import { getCaseFile, upsertNote } from "../../../../../store/reducers/cases";
+import { upsertNote, getCaseFile } from "../../../../../store/reducers/case-thunks";
 
 type props = {
   id: string;
