+++ frontend/src/app/components/containers/complaints/outcomes/supplemental-notes/supplemental-notes-item.tsx
@@ -5,7 +5,7 @@ import { useAppSelector } from "../../../../../hooks/hooks";
 import { formatDate } from "../../../../../common/methods";
 
 import { CaseAction } from "../../../../../types/outcomes/case-action";
-import { selectNotesOfficer } from "../../../../../store/reducers/cases";
+import { selectNotesOfficer } from "../../../../../store/reducers/case-selectors";
 
 type props = {
   notes: string;
